Migrate to new slice API

+++ librubyfmt/src/types.rs
@@ -1,4 +1,4 @@
-pub type LineNumber = u64;
-pub type SourceOffset = usize;
+pub type LineNumber = i32;
+pub type SourceOffset = u32;
 // 4 billion characters ought to be enough for anyone
 pub type ColNumber = u32;
